requirements.txt
requests
python-binance